+++ src/App.tsx
@@ -8,7 +8,7 @@ import 'react-toastify/dist/ReactToastify.css'
 import IndexingProgress from './components/Common/IndexingProgress'
 import MainPageComponent from './components/MainPage'
 import InitialSetupSinglePage from './components/Settings/InitialSettingsSinglePage'
-import { ModalProvider } from './providers/ModalProvider'
+import { ModalProvider } from './contexts/ModalContext'
 
 interface AppProps {}
