+++ src/components/TitleBar/TitleBar.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import { PiSidebar, PiSidebarFill } from 'react-icons/pi'
 import DraggableTabs from '../Tabs/TabBar'
-import { ModalProvider } from '../../providers/ModalProvider'
+import { ModalProvider } from '../../contexts/ModalContext'
 import FileHistoryNavigator from './NavigationButtons'
 
 export const titleBarHeight = '30px'
